Задание четвертое. Исправила замечания по модулям
