+++ js/main.js
@@ -1,3 +1,5 @@
-import './data';
+import {createAds} from './data';
+const SIMILAR_OBJECT_COUNT = 10;
 
+createAds(SIMILAR_OBJECT_COUNT);
